% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPDBs.R
\name{getPDBs}
\alias{getPDBs}
\title{Pattern Discovery in PDB Structures of Metalloproteins}
\usage{
PdPDB(path, metal, n, perc, interactive, dropsReplicate)
}
\arguments{
\item{PDBlist}{A vector of strings. Contains the target PDB files to be downloaded.}

\item{path}{A string containing the local path to which the PDBs will be downloaded.}

\item{ext}{A string. Either 'pdb' or 'cif'.}
}
\description{
Download a list of PDB files to a local path
}
\note{
Files will be downloaded in a local folder as specified in path.
}
\references{
Belmonte L, Mansy SS Patterns of Ligands Coordinated to Metallocofactors Extracted from the Protein Data Bank, Journal of Chemical Information and Modeling (accepted)
}
\author{
Luca Belmonte
}
\keyword{PDB,}
\keyword{alignment,}
\keyword{coordinating}
\keyword{ligand}
\keyword{metal,}
\keyword{metalloproteins,}
\keyword{motifs}
\keyword{patterns,}
